+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # JEC JARS Project: Update Release Notes
 
+<a name="jec-jars-1.0.8"></a>
+## **1.0.8** (2017-12-26)
+
+### Bug Fixes
+
+### Features
+
+- Dependencies upgrade
+- Fixing `jec-commons` module break changes
+
 <a name="jec-jars-1.0.7"></a>
 ## **1.0.7** (2017-09-07)